Require modal__close-icon class for close icon
The class modal__close should be available for any element in the
modal indicating that clicking closes the modal but not necessarily
that this element should be styled as a close icon.
The close icon now needs the classes modal__close and
modal__close-icon, where modal__close takes care of the
functionality and modal__close-icon takes care of the styling.