Fix escape out of focus-trap
When opening a modal the modal is selected. Pressing shift+tab let
you escape the modal, as the event listeners for shift+tab and tab
only checked on the first and last focusable elements.
Now the event listeners also listen on the modal.
Additionally, the first and last focusable elements default to the
modal itself incase no focusable element exists.